+++ ChangeLog.md
@@ -1,2 +1,2 @@
-﻿- Potential bugfix for freezing keyboard when playing in windowed mode
-- Some refactoring and cleanup of the codebase
+﻿- Added brush size and strength scrollwheel actions for tree brush
+- Fixed minor typo in the settings menu